IS_ontology/IS_ontology/api/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
    TokenVerifyView,
)

from .views import get_terms, get_triples, user_info


app_name = "api"

router = DefaultRouter()

urlpatterns = [
    path("terms/", get_terms),
    path("triples/", get_triples),
    path("token/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
    path("token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
    path("token/verify/", TokenVerifyView.as_view(), name="token_verify"),
    path("user/", user_info)
]
