IS_ontology/requirements.txt
absl-py==1.4.0 ; python_version >= "3.10" and python_version < "3.11"
asgiref==3.7.2 ; python_version >= "3.10" and python_version < "3.11"
astunparse==1.6.3 ; python_version >= "3.10" and python_version < "3.11"
beautifulsoup4==4.12.2 ; python_version >= "3.10" and python_version < "3.11"
black==23.7.0 ; python_version >= "3.10" and python_version < "3.11"
cachetools==5.3.1 ; python_version >= "3.10" and python_version < "3.11"
certifi==2023.7.22 ; python_version >= "3.10" and python_version < "3.11"
cffi==1.15.1 ; python_version >= "3.10" and python_version < "3.11"
charset-normalizer==3.2.0 ; python_version >= "3.10" and python_version < "3.11"
click==8.1.6 ; python_version >= "3.10" and python_version < "3.11"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.11" and platform_system == "Windows"
cryptography==41.0.3 ; python_version >= "3.10" and python_version < "3.11"
dawg-python==0.7.2 ; python_version >= "3.10" and python_version < "3.11"
django-bootstrap4==23.2 ; python_version >= "3.10" and python_version < "3.11"
django-cors-headers==4.2.0 ; python_version >= "3.10" and python_version < "3.11"
django-debug-toolbar==4.2.0 ; python_version >= "3.10" and python_version < "3.11"
django-htmx==1.16.0 ; python_version >= "3.10" and python_version < "3.11"
django==4.2.4 ; python_version >= "3.10" and python_version < "3.11"
djangorestframework-simplejwt[crypto]==5.2.2 ; python_version >= "3.10" and python_version < "3.11"
djangorestframework==3.14.0 ; python_version >= "3.10" and python_version < "3.11"
docopt==0.6.2 ; python_version >= "3.10" and python_version < "3.11"
flatbuffers==23.5.26 ; python_version >= "3.10" and python_version < "3.11"
gast==0.4.0 ; python_version >= "3.10" and python_version < "3.11"
google-auth-oauthlib==1.0.0 ; python_version >= "3.10" and python_version < "3.11"
google-auth==2.22.0 ; python_version >= "3.10" and python_version < "3.11"
google-pasta==0.2.0 ; python_version >= "3.10" and python_version < "3.11"
grpcio==1.57.0 ; python_version >= "3.10" and python_version < "3.11"
gunicorn==20.1.0 ; python_version >= "3.10" and python_version < "3.11"
h5py==3.9.0 ; python_version >= "3.10" and python_version < "3.11"
idna==3.4 ; python_version >= "3.10" and python_version < "3.11"
joblib==1.3.2 ; python_version >= "3.10" and python_version < "3.11"
keras-preprocessing==1.1.2 ; python_version >= "3.10" and python_version < "3.11"
keras==2.13.1 ; python_version >= "3.10" and python_version < "3.11"
libclang==16.0.6 ; python_version >= "3.10" and python_version < "3.11"
markdown==3.4.4 ; python_version >= "3.10" and python_version < "3.11"
markupsafe==2.1.3 ; python_version >= "3.10" and python_version < "3.11"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "3.11"
neo4j==5.11.0 ; python_version >= "3.10" and python_version < "3.11"
nltk==3.8.1 ; python_version >= "3.10" and python_version < "3.11"
numpy==1.24.3 ; python_version >= "3.10" and python_version < "3.11"
oauthlib==3.2.2 ; python_version >= "3.10" and python_version < "3.11"
opt-einsum==3.3.0 ; python_version >= "3.10" and python_version < "3.11"
packaging==23.1 ; python_version >= "3.10" and python_version < "3.11"
pathspec==0.11.2 ; python_version >= "3.10" and python_version < "3.11"
platformdirs==3.10.0 ; python_version >= "3.10" and python_version < "3.11"
protobuf==4.24.0 ; python_version >= "3.10" and python_version < "3.11"
psycopg2-binary==2.9.7 ; python_version >= "3.10" and python_version < "3.11"
pyasn1-modules==0.3.0 ; python_version >= "3.10" and python_version < "3.11"
pyasn1==0.5.0 ; python_version >= "3.10" and python_version < "3.11"
pycparser==2.21 ; python_version >= "3.10" and python_version < "3.11"
pyjwt==2.8.0 ; python_version >= "3.10" and python_version < "3.11"
pymorphy2-dicts-ru==2.4.417127.4579844 ; python_version >= "3.10" and python_version < "3.11"
pymorphy2==0.9.1 ; python_version >= "3.10" and python_version < "3.11"
python-crfsuite==0.9.9 ; python_version >= "3.10" and python_version < "3.11"
pytz==2023.3 ; python_version >= "3.10" and python_version < "3.11"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "3.11"
regex==2023.8.8 ; python_version >= "3.10" and python_version < "3.11"
requests-oauthlib==1.3.1 ; python_version >= "3.10" and python_version < "3.11"
requests==2.31.0 ; python_version >= "3.10" and python_version < "3.11"
rsa==4.9 ; python_version >= "3.10" and python_version < "3.11"
setuptools==68.0.0 ; python_version >= "3.10" and python_version < "3.11"
six==1.16.0 ; python_version >= "3.10" and python_version < "3.11"
sklearn-crfsuite==0.3.6 ; python_version >= "3.10" and python_version < "3.11"
soupsieve==2.4.1 ; python_version >= "3.10" and python_version < "3.11"
sqlparse==0.4.4 ; python_version >= "3.10" and python_version < "3.11"
tabulate==0.9.0 ; python_version >= "3.10" and python_version < "3.11"
tensorboard-data-server==0.7.1 ; python_version >= "3.10" and python_version < "3.11"
tensorboard==2.13.0 ; python_version >= "3.10" and python_version < "3.11"
tensorflow-estimator==2.13.0 ; python_version >= "3.10" and python_version < "3.11"
tensorflow-io-gcs-filesystem==0.31.0 ; python_version >= "3.10" and python_version < "3.11"
tensorflow==2.13.0 ; python_version >= "3.10" and python_version < "3.11"
termcolor==2.3.0 ; python_version >= "3.10" and python_version < "3.11"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "3.11"
tqdm==4.66.1 ; python_version >= "3.10" and python_version < "3.11"
typing-extensions==4.5.0 ; python_version >= "3.10" and python_version < "3.11"
tzdata==2023.3 ; python_version >= "3.10" and python_version < "3.11" and sys_platform == "win32"
uritemplate==4.1.1 ; python_version >= "3.10" and python_version < "3.11"
urllib3==1.26.16 ; python_version >= "3.10" and python_version < "3.11"
werkzeug==2.3.6 ; python_version >= "3.10" and python_version < "3.11"
wheel==0.41.1 ; python_version >= "3.10" and python_version < "3.11"
wrapt==1.15.0 ; python_version >= "3.10" and python_version < "3.11"
